Calm down the ANF transformation that precedes the autobatching transformation.
This should increase the amount of Python syntax that "just works"
inside autobatched functions.

Specifically, the autobatching transform used to just convert all of
the batchee to A-normal form.  Now that the underlying A-normal form
converter is configurable, the autobatching transform instead just
insists on fresh variables being introduced for the conditions of `if`
statements, for all function call expressions, and for the arguments
to auto-batched functions.  Those are the places the down-stream
source transformation needs to think about.

PiperOrigin-RevId: 261954934