Add `tfp.vi.build_factored_surrogate_posterior` utility for automatic black-box variational inference.
This is a step towards the vision of go/tfp-awesome-vi. Eventually we will also
want utilities to build variational distributions that have dependence across
model variables (MVNs, IAFs, etc), but this will require first landing go/multipart-bijectors or some alternative. For now, we can enable the common case of
mean-field VI by explicitly building JDs out of component parts.

Q: Is this just a stopgap, or does it have a long-term future?
A: We will eventually implement a more general utility that applies a trainable
bijector to a base Normal distribution, then splits the result. This will in
principle allow users to replicate the functionality of
`build_factored_surrogate_posterior` (by plugging in an AffineScalar bijector).
In practice, I expect users who want factored posteriors
will continue to find the explicit formulation in this CL more natural, so
I think this CL will be useful even once we have more general solutions.

Q: This method supports either of `event_shape` or `initial_loc` to specify the
   structure of the posterior. Why do we need both?
A: It's clear we need `event_shape`: it's the minimal contract, it's often more
   convenient for users (since we do the work of generating the initial loc for
   them), and it avoids potentially unnecessary bijector inverses.
   Supporting `initial_loc` as an alternative is not *essential* but IMHO it's
   nice to have: it gives users control over initialization, which is often
   valuable, and it also directly matches the MCMC API.
PiperOrigin-RevId: 268745721