Avoid testing Von Mises sample moments in low concentration regime.
The (circular) sample mean estimator has fairly high variance for low
concentration values. This means the test looking at sample moments is pretty
flaky if we vary the seed (failing ~30% of the time). Previously it just had a
hardcoded seed. When enabling different test modes in cl/269868487, that seed
lost its magical powers and the test became flaky.

Therefore, this CL eliminates the test case where the concentration is 0.1 and
gets rid of the magic seed.

PiperOrigin-RevId: 274295088