Support posterior sampling in linear Gaussian state space models.
Also now includes some minor cleanup of shape broadcasting.

This implements the Durbin-Koopman algorithm, which samples from the posterior without explicitly realizing its (giant, intractable) joint covariance. It exploits the fact that the posterior covariance in Gaussian models is independent of the observed values. In particular, we draw a zero-mean sample of the latents z with the appropriate covariance by
1. Drawing a *prior* sample z, x ~ p(z)p(x|z)
2. Computing the posterior mean u = E[z | x].
3. Proving that the quantity z - u has mean zero and the same covariance as a posterior sample. [1]

[1] http://www.datascienceassn.org/sites/default/files/A%20Simple%20and%20Efficient%20Simulation%20Smoother%20for%20State%20Space%20Time%20Series%20Analysis.pdf

PiperOrigin-RevId: 274696272