Eliminate make_(diag|tril)_scale from distribution_util. Many consequences.
These methods were contributing to anti-patterns of doing work in init, and in
a manner which would have been quite hard to eliminate without a big mess. The
things that crucially relied on them are deprecated by this change, and in the
mean time the make*scale functions are duplicated locally in the deprecated
modules. These can all be wiped at the end of the deprecation window.

Bulleted summary of changes herein:
  - Deprecate several distributions (which would have been hard to make tape safe and for which better alternatives exist):
    - MVNDiagPlusLowRank (likely no one really wants this anyway, see #161)
    - Vector{Laplace,Exponential,SinhArcsinh}Diag -- more efficient to use
      tfd.Independent.
  - Delete (private, seemingly unused) _VectorStudentT distribution
  - Deprecate MVNDiag's `scale_identity_multiplier` argument. Explicitly
    disallow mixing it with `scale_diag` argument.
  - Make MVNDiag tape-safe; it previously wasn't completely so, since
    scale_diag and scale_identity_multiplier could sometimes be added during
    init.

PiperOrigin-RevId: 274954759