Decrease likelihood of test timeouts for slice sampler test.
Note: this makes the tolerance on eager uncomfortably high; we should prioritize solving the performance issue in some way, or push these tests off to (WIP) nightly test setup.

1. dramatically reduce number of samples drawn and numbers of parallel chains
2. make graph and eager mode use different numbers of samples and tolerances

Also:

1. Go ahead an decorate with new test decorator. This helps a bit by allowing
graph and eager modes to run in separate test shards. It also helps ensure
we're not going to be subject to further flakes when the new test hardness is
spun up (soon!)

2. refactor nearly-identical tests, using absl.testing.parameterized
3. simplify existing tests, a bit, which were using list-like initial state
parts when the didn't need to. added a single, separate test to exercise
list-like initial state parts.

Will separate investigate a bit into why eager is mode is *so* much slower.

PiperOrigin-RevId: 276512850