Remove `_most_recently_built_distributions` from JointDistributions, and make batch shape a function of sample shape.
This change makes explicit the existing semantics, in which the (ragged) batch shape of
a joint distribution depends on the passed-in sample shape. It adds a `sample_shape` argument to `batch_shape_tensor`, set to `()` by default in order to match the behavior of the `batch_shape` property (which cannot take an argument).

The ability to explicitly query batch shape given a sample shape removes the need for `_most_recently_built_distributions`, so we replace it with `_single_sample_distributions`, which is equivalent except that it always represents a draw with sample shape `()`. This has minimal effect on the semantics of existing JDs, but will make it easier to implement new JD semantics in subclasses.

Unlike `_most_recently_built_distributions`, `_single_sample_distributions` is instantiated on first use, even upon accessing a property. This guarantees that distribution properties are always well-defined. This deviates from the principle that properties shouldn't have graph side effects, but we believe this is justified because:
- JointDistributions present unique challenges, and ensuring that properties are well-defined is also an important design goal.
- The graph side effect occurs only on first access to any property (or method); future accesses have no side effects because they hit the cache.
- The TF team has committed to making graph construction faster.

As a bonus, we add logic to invalidate the cached distributions if the graph context has changed.

Finally, this change does away with returning the distributions from `_map_attr_over_dists` and `map_measure_over_dists`. This simplifies the code (since they are never used), and makes it easier to write drop-in replacements for those methods that use vectorized_map.

PiperOrigin-RevId: 277782203