Use the known `dtype` structure of JDs to intelligently convert values to Tensor.
Just as JDs have the information in `self.dtype` to interpret args and kwargs, they also have the information to apply `convert_to_tensor` at exactly the right nesting level. This is necessary in general to avoid problems when passing nested lists as JD values.

As a bonus, this lets us delete the specialized JD implementation of `_call_log_prob`, because the base Distribution class now handles structured dtypes properly.

PiperOrigin-RevId: 281852328