Fix memory leak in bijector cache (for py3).
The root cause here was a combination of a few (very subtle) things.

  1. HashableWeakRef.__hash__ computed the hash of the referrent, which may
     have died and been cleaned up before the HWR instance's cleanup callback
     was called.
  2. WeakKeyDefaultDict initialized HashableWeakRef instances with a cleanup
     callback that *tried* to remove the item from the dict, using pop, but did
     so with a default value (ie, dict.pop(w, None)) which masked the fact that
     this was frequently failing. This pop call depends on the hash of the HWR
     instance, which in turn needed to be computed from the referrent, which
     was (only sometimes!) already dead and gone.

The fix that this CL proposes is to mimic what weakref.ref does in its own hash
routines (see
https://github.com/python/cpython/blob/master/Objects/weakrefobject.c#L25 and
https://github.com/python/cpython/blob/master/Objects/weakrefobject.c#L146).

Namely, we memoize the hash value, which ensures that even if the referrent has
been destroyed, we can still locate the HWR instance in the dict, and clean it
up.

Fixes: #647
PiperOrigin-RevId: 282013709