Speed up ode_test by not testing automatic computation of Jacobians with pfor in Eager.
Why?  When the ambient solve is being done in Eager mode, pfor ends up
being invoked ~500 times to construct a tf.function for the Jacobian,
which dominates the runtime.

This change speeds this target up from a latency of ~400 seconds to
~100 seconds, and eliminates the large previous non-uniformity in
per-method test runtimes.  The total CPU time of the target falls from
1000s to 360s.

We are still testing the ODE solver in Eager mode, on several targets
with user-provided Jacobians of the system of equations.

PiperOrigin-RevId: 284766078