FunMCMC: Improve trace API.
1. Over the past few months of usage, it became clear that a configurable
`trace_fn` is not necessary and only adds to the boilerplate. It'll be removed
in the future, but for now I added a default implementation that traces the
`extra` output.

2. One perennial annoyance is that it was hard to get the final `extra` output
without tracing it for all timesteps. The `trace_mask` addresses this issue.
Now it's trivial to do MCMC chain thinning via composition:

```
def hmc(s):
  s, e = fun_mcmc.hamiltonian_monte_carlo(s, ...)
  return s, s.state

fun_mcmc.trace(
    fun_mcmc.hamiltonian_monte_carlo_init(...),
    lambda s: fun_mcmc.trace(
        s,
        hmc,
        num_steps=num_substeps,
        trace_mask=False,
    ),
    num_steps=num_steps // num_substeps,
)
```

PiperOrigin-RevId: 285072496