FunMCMC: Add explicit API for computing exponential moving averages.
This was already possible before via some hacking, but this makes it more accessible.

Also, fix a few incorrect nesting code and typos.

A comment on the debiasing technique here and how it relates to what's done by Adam:

Both approaches are unbiased, but perform something different in the transient phase at the beginning of averaging. The method here has lower variance, but arguably isn't really doing EMA like Adam's debiasing does.

I argue that there is no obvious preference of one approach over the other in the transient phase, so this seems a reasonable API for a person who is interested in a debiased EMA.

PiperOrigin-RevId: 285284795