Avoid slow tf.nest.unpack_sequence_as if its result is not needed.
This place is showing up in Python profiles across all our slowest
Eager-mode tests (which are generally characterized by long loops over
many small-scale ops).  The obvious change to just skip everything
when in Eager isn't correct, because of the possibility of Keras
symbolic Tensors (per the comment), leaving this small improvement.
Ironically, the improvement is expected to improve graph construction
rather than Eager execution in the common case, but it does seem to
help some.

For example, mcmc:transformed_kernel_test.python3 (100 independent
trials, sharded 4 ways each) speeds up from
  Stats over 400 runs: max = 304.6s, min = 81.4s, avg = 147.1s, dev = 35.0s
to
  Stats over 400 runs: max = 257.8s, min = 67.0s, avg = 138.3s, dev = 35.9s

This test is a relatively extreme user of prefer_static, at least in
Eager mode, and of course there is room for noise in this measurement,
but still.

PiperOrigin-RevId: 285561549