Avoid testing sanity of Generalized Pareto log-prob at extreme location-scale pairs.
Why?  This distribution is a location-scale family, whose sampler and
log-prob (and log-cdf) are implemented with the obvious transforms `x
= loc + scale * z` and `z = x / scale - loc`, for `x` the sample of
interest as `z` a standardized sample.  When `loc / scale` approaches
`1 / eps` (i.e., around 1e8 in 32-bit floats), these two
transformations incur catastrophic loss of precision, to the point
that post-transform `z` can register as outside the support.

PiperOrigin-RevId: 285561591