FunMCMC: Refactor the `trace` operator.
The goal was not to have backend-conditional logic in the main file. A side
benefit is that by using JAX's excellent 'scan' function, we also gain the
ability to differentiate through the computation. In general, this is a big
improvement to the JAX usability and speed of this code.

Along the way I removed the 'None' imputation code. It was a cute idea, but I
never ended up using it and none of the more recent code in FunMCMC supports
that convention anyway. It just isn't that hard to just call the `_init`
method, and the rest of the imputation needs are well handled by the `extra`
return.

PiperOrigin-RevId: 288954947