FunMCMC: Add a new reparameterization API.
The idea of this API is to move away from the world of bijectors, towards a
world where invertible transformations are created by writing regular functions and then using function transformations (e.g. via JAX's interception functionality). In this world, we reparameterize using functions (which, in FunMCMC land, also return extra
output) which can be inverted and have their LDJs computed automatically.

One design decision is how to handle event_ndims. In bijectors, we must do a
final reduction based on a per-event event_ndims. In this API, we always assume
that's fixed. This seems less of a problem than it was before now that we have
vmap. In the vmap world it seems okay to restrict ourselves to always operating on one event at a time and then rely on vmap to handle the extraneous dimensions. For convenience with the current world, however, I still added a `batch_ndims` argument.

PiperOrigin-RevId: 288966593