Define an interface for convergence criteria, and add a proof of concept example.
This is the first part of a two-part change adding support for custom convergence criteria to `tfp.math.minimize`. The second change cl/286485608 will incorporate convergence criteria into the training loop. (it's actually already 'finished'/ready for review, but I split off this piece to make the reviewing a little easier and allow us to consider the interface in isolation).

The `ConvergenceCriterion` base class defines an interface for convergence criteria somewhat inspired by MCMC transition kernels. A convergence criterion implements two methods:

  - a method `has_converged, new_state =
                _one_step(loss, grads, values, state)`
    that takes the current loss, grads, and values of parameters being optimized,
    as well as a criterion-specific state that may preserve information from
    previous steps, e.g., a moving average of the loss, parameters, gradients,
    etc. It returns whether
    the optimization should stop at this step, as well as an updated state.

  - A method `_initial_state(loss, grads, values)`, which determines the (potentially
    nested) structure of the state by initializing a state from the initial loss, grads,
    and values. E.g., a moving average of loss might be initialized to the initial loss.

As part of this change, we add a simple convergence criterion `LossDecrease`, which stops when the change in a moving average of loss falls below an absolute or relative (to the loss decrease so far) tolerance, as proof of concept. As part of developing the interface I've also prototyped a couple of additional convergence criteria, in cl/286675392 and cl/286675325, which I hope to submit later once the interface is polished.

PiperOrigin-RevId: 292197531