FunMCMC: Add Simple Dual Averaging implementation.
As far as I can tell, this algorithm is implemented correctly (cross checked with publications as well as tfp.mcmc.DualAveragingSSA), but it's pretty clear that it's extremely sensitive to the Lipschitz constant of the function being optimized, and is *extremely* slow to converge. It probably makes sense for baselines, but I wouldn't ever reach for it over Adam or other modern subgradient descent algorithms.

PiperOrigin-RevId: 293081652