Adjust default "hmc-like" step size getter and setter functions to work with NUTS (as well as HMC).
The problem is that HMC is composed of MH and uncalibrated HMC, so it
exposes its step_size parameter as
kernel_results.accepted_results.step_size, whereas NUTS exposes it
directly as kernel_results.step_size.

This was preventing our own step size adaptation kernels from working
with our own NUTS kernel out of the box.

This CL consists of the band-aid, pending a more reasoned decision about how to adjust the API.

PiperOrigin-RevId: 294309754