Support dynamic `num_steps` and custom convergence criteria in `tfp.math.minimize`.
This CL adds a `convergence_criterion` argument to `tfp.math.minimize`. The default of `None` preserves existing behavior. The user now also has the option to pass an instance of `tfp.optimizer.convergence_criteria.ConvergenceCriterion`, specifying a custom policy for stopping the optimization. In this case, the optimization will run for a dynamic number of steps, up to a maximum of the specified `num_steps`. As a bonus, `minimize` now also supports dynamic `Tensor` values for `num_steps`.

The state propagated by the convergence criterion is not traced by default. Instead, we update
the `trace_fn` to take a MinimizerState namedtuple containing the optimization state as well as the state of any convergence criterion, so that users can decide for themselves
which aspects of the state to trace.

I anticipate propagating the `convergence_criterion` argument at least to `tfp.vi.fit_surrogate_posterior` (which wraps `tfp.math.minimize`) in a future change.

PiperOrigin-RevId: 300228821