FunMC: Add Adaptive HMC prefab.
FunMC makes it easy to write new MC transition operators, but nothing is easier
than just importing something pre-made. The prefab module addresses this use
case.

This is also a beginning of Phase 2 of Inference Gym effort: a "zoo" of
inference algorithms that we'll verify (in terms of efficiency and robustness)
against a multitude of targets.

It also serves a purpose of "I just want to do inference, not do MCMC research
from scratch!" need: aka "turn-key" MCMC inference. Right now
`adaptive_hamiltonian_monte_carlo` still takes a bit of boilerplate to get
started, so perhaps there might be more simplification in the future.

PiperOrigin-RevId: 301678787