Allow using the adjoint sensitivites method to compute gradients with respect to values constant wrt time.
Previously, this worked either by using tf.Variables for these values or
by augmenting the state (with corresponding zero gradients). The latter solution
was inefficient, as it wasted computation for the forward pass of these
constants.

Now, you can specify these constants with a `constants` argument which is
passed into the `ode_fn` as kwargs. In a toy model, this led to ~25% speedup.

PiperOrigin-RevId: 303063766