Add `infer_trajectories` utility wrapper around particle filtering.
This is motivated by a forthcoming change introducing adaptive resampling, which will cause the particle filter to generate weighted particles at some steps. This is statistically helpful but kind of a pain to deal with on the user side.

My proposal is that we bite the bullet and allow `particle_filter` to return stuff like weights and indices, for power users, but also expose a convenience wrapper that just spits out unbiased samples of trajectories for users who don't need the jucier bits.

PiperOrigin-RevId: 304719704