Support custom tracing of particle filter results.
This adds two new arguments to `particle_filter`:
 - `trace_fn` specifies *what* to trace (e.g., the full particles, weights, and indices, or maybe just the marginal likelihoods)
 - `trace_only_at_steps` specifies *when* to trace. (e.g., at every step, on the 2nd, 5th, and 29th steps, or just at the last step).
Note that `infer_trajectories` does not expose these arguments, because it already relies on a particular trace structure to extract the trajectories.

This is in preparation for a couple of other upcoming changes:
 - Adding an optional `extra` field propagated between steps, to match the interface of Srinivas's EnKF. This change will decouple that change (and others in the future) from the user-facing API.
 - Implementations of particle MCMC and SMC**2, both of which run an inner-loop particle filter to compute an approximate marginal likelihood, but don't care about the
particles at all. This change will let us avoid accumulating the particle history in memory when it's not needed.

PiperOrigin-RevId: 309096556