Use ascending sort for percentile
This changes the behavior of percentile and quantiles, as previously the
sort order had been descending. This is being changed as it differs in
behavior from NumPy's percentile function and the docstring describes
behavior that is consistent with NumPy's implementation.

The main change occurs in _sort_tensor, which is then used in
percentile. In _sort_tensor tf.math.top_k (which results in descending
sort) is swapped for tf.sort (with the default values of k=-1,
direction='ASCENDING' explicitly passed to make the behavior more
explicitly clear to a reader). This then necessitates the change in
definition to using frac_at_q_or_below, and the swap of tf.math.ceil and
tf.math.floor along with 'higher' and 'lower'.

Also break up the docstring for _sort_tensor to make the linter happy.