Add test for edge case
Add a test that checks the edge case that resulted in the PR. The case
is that of the NumPy percentile docs example using the "nearest"
interpolation method:

a = np.array([[10, 7, 4], [3, 2, 1]])
np.percentile(a, 50, interpolation="nearest")

NumPy returns 3, but before this PR TFP's result was 4. This test
validates this edge case (both return 3).