Refactor SMC and particle filtering to share resampling methods.
This change updates `sample_sequential_monte_carlo` to share resampling code with particle filtering, including Dan's new fast samplers. As part of this, it deletes the particle filter's `batch_gather` in favor of `mcmc_util.index_remapping_gather`.

It also adds a `resample_fn` argument to the relevant methods, allowing the user to specify the resampling method of their choice.

In order to pass the SMC tests, which use XLA, I had to remove use of the built-in tf.math.cumulative_logsumexp. Rather than use a pure TF implementation as a workaround, which might be significantly slower, I've updated the samplers to do their calculations in linear space using `tf.cumsum`. Our hypothesis (which was discussed in cl/305882181) is that this may have numerical properties as good or better than the log-space implementation.

PiperOrigin-RevId: 312174475