Factor `SequentialMonteCarlo` transition kernel out of the particle filter.
cgs@ convinced me to take advantage of the loop driver infrastructure that already exists for MCMC, instead of reinventing a special-purpose SMC driver. This allows the 'SMC' implementation to be just a TransitionKernel.

This CL also removes the SampleParticles distribution, which is no longer needed by the particle filter because it can now just pass initial particles directly into the SMC kernel.

PiperOrigin-RevId: 314605953