FunMC: Simplify and stabilize running statistics.
The auto-covariance likely can be simplified in the same way, but it's more
complicated so I'm punting it for now.

Also, initialize variances with 1's. This doesn't affect the outputs past the
first update step, but is pretty convenient when scaling anything by those
statistics, e.g. in the context of adaptive MCMC.

PiperOrigin-RevId: 317721306