Adds seed argument to `TransitionKernel.one_step`, and `tfp.mcmc.sample_chain`.
Deprecates passing of seeds to `TransitionKernel.__init__`.

As with `tfd.Distribution.sample`, a seed is sanitized by `sample_chain`:
- from an int to a `Tensor` seed pair, using `tf.random.uniform`, yielding the already existing (somewhat surprising in eager) TF stateful sampling behavior
- incoming seed pairs are preserved as such, and will yield stateless (completely deterministic) sampling behavior.

The built-in TransitionKernel instances are modified to trace their seeds in kernel results namedtuples. This can be useful, e.g. for debugging a divergence. The MH kernel replaces fields named 'seed' in the 'accepted' portion of the kernel results with empty lists, since these fields generally come from either the proposed or previously accepted result.

PiperOrigin-RevId: 317945519