Use a custom build rule to export symlinks to generated jax/numpy implementations under `tensorflow_probability/substrates/{jax,numpy}`. This way we can benefit from existing test coverage, but avoid TF dependencies that come from `__init__.py` files under `tensorflow_probability/python/`, which were previously being executed when importing (e.g. `tensorflow_probability.python.distributions._jax`).
This also amends the pip package structure so that TFP on JAX lives under tensorflow_probability/substrates instead of in nested _jax subdirectories under the corresponding TF directories.

At build time, files are still generated to a similar location (e.g. a _jax or _numpy subdirectory of the ambient rule), but with a _generated_ prefix. This ensures that when building the wheel later, it will not find a __init__.py and will omit these directories.

PiperOrigin-RevId: 322648215