Add stochastic volatility model to inference gym.
This is the same model described in the Stan user manual:
https://mc-stan.org/docs/2_21/stan-users-guide/stochastic-volatility-models.html
a related stochastic volatility model is featured in the NUTS paper
https://arxiv.org/abs/1111.4246.

As a dataset, I retrieved closing prices from the S&P 500 over the past ten years. My understanding is that this information is considered factual and therefore not covered by copyright.

PiperOrigin-RevId: 322841441