Make DeferredTensor actually defer computation under JAX/NumPy backends.
While historically the primary usage of DeferredTensor has been for tape safety, something only of concern in TensorFlow, there are other potentially valid use cases for deferring computation.

Aside from minor papercuts, the main implementation challenge was implementing operator overloading. This was made tricky by the fact that JAX uses different types for its array objects depending on whether computation is being traced or not. This necessitated making the concretization logic be deferred until we knew the type of the return value of transform_fn.

There are a number of other oddities, dealing with __array_priority__, __nonzero__ and __bool__ which I don't fully understand, but the tests pass, so perhaps it's all good.

A number of tests were disabled as they rely on finer details of tf.Module and tf.Variable which the other backends do not implement.

A number of places in the numpy backend got new `convert_to_tensor` calls, which is a consequence of jax not doing them itself and balking at the DeferredTensor passed to those functions.

PiperOrigin-RevId: 323505584