Change the naming scheme of un-named variables in JointDistributions.
1. The names are now valid Python identifiers, composed of the prefix `_var` and a counter suffix.
2. The counter is now corresponds to the index of the RV in the tuple representation of the event shape (i.e. it's zero indexed, and is incremented at each RV site).

(1) is done to enable us to use namedtuples are return dtype for JointDistributionCoroutine in the future.
(2) is done in hope of making it slightly easier to predict the name of the automatically named variable.

This is a breaking change, but the predicted scope is so low, we're doing this without a deprecation period.

PiperOrigin-RevId: 323870754