Defer loading datasets until computing the likelihood.
This helps cases where we just want to get the name etc from the model.

Two models were not adjusted appropriately:

ItemResponseTheory: This is blocked on JAX/NumPy backend wrapping scatter_nd, but is otherwise perfectly doable.

LogGaussianCoxProcess: This requires a separate approach since the "prior" is data dependent (cf the inducing_points on the GP prior). We can still do it by utilizing a tracing context, but that requires a bit of thinking and is best deferred to a separate CL.

One issue shared by both models is that they construct the prior conditioned on some data processing in the initializer (extracting number of student/question ids and extracting the mean log intensity respectively). Those will be worked around by placing that information into the datasets themselves.

PiperOrigin-RevId: 323915226