Support vector-valued offsets in sts.Sum.
This patches an annoying hole in the current API, noted in
https://groups.google.com/a/tensorflow.org/g/tfprobability/c/b0Yf0DNOlZY/m/3IHzc9nWBAAJ

This is technically a breaking change, because a vector-valued `constant_offset` is now interpreted as time-varying, when previously it would have been a batch of scalar offsets. However, I suspect that the number of people who are both:
 - using STS with batch inputs, and
 - manually setting the constant_offset parameter to a batch value (as opposed to passing a batched `observed_time_series` which will set this programmatically)
is close enough to zero that this is not worth a deprecation cycle.

PiperOrigin-RevId: 326083693