Fix shape and dtype errors to enable ground truth checks for stochastic volatility model.
Enabling `validate_log_prob_and_transforms` ground truth checks in `stochastic_volatility_test` surfaced two issues.

First, the dtype for the `identity` sample transformation was being incorrectly set to the default of `tf.float32` instead of a structure. This led to a silent failure to load the ground truth.

Second, the scalar RVs were being incorrectly extracted from Stan and saved as length-1 vectors in the ground truth file. This caused shape errors when comparing them to the true scalars in the TF version.

This change fixes both issues.

PiperOrigin-RevId: 327549821