Inference Gym: Move Inference Gym to the spinoffs directory.
This does 3 things simultaneously:

1. Conversion to Python 3-only.
2. Switch to the dynamic backend rewrite system.
3. The actual move.

(2) implies (1), and (3) is easier with (2) (as otherwise we'd need to rejiggle the static rewrite system, which would be wasted work if doing (2) is already assured in the future).

This is likely not the final state of things (cf what Oryx does), but it accomplishes a significant level of separation, enabling the early Python 3 transition.

PiperOrigin-RevId: 327918957