Handle `nan`s in the rejection sampling utility.
To wit, previously `nan`s would cause the rejection sampler to reject.
This is semantically bad, but especially manifests in the common case
that the `nan` occurs on every proposal (e.g., because some parameter
was `nan`), where it becomes an infinite loop.

Now, if a `nan` occurs, we force-accept and return a `nan` result in
that batch member.  Detailed rationale in code comment.

PiperOrigin-RevId: 328191523