Inference Gym: Add a cludge to enable `using_jax/numpy` to actually work without TensorFlow.
The issue is that when using `find_spec('foo.bar')`, a core operation in the dynamic rewriting system, Python will execute
`foo/__init__.py`. Normally this isn't an issue, but `targets/__init__.py`
actually re-exports API, causing all of the platform specific dependencies
(e.g. TensorFlow) to be imported. This raises an error if TensorFlow isn't linked in (which is something we want to allow when using JAX/NumPy backends).

This cludge disables those import errors if they are found in the unrewritten
`__init__.py` files. If a legitimate import error is present, it'll be raised
inside the rewritten modules instead.

Added some integration tests to verify that this all works (they may seem a
little circular, but they legitimately failed before this change).

PiperOrigin-RevId: 328403005