Test that BetaBinomial samples correctly and doesn't enter an infinite loop, even under XLA.
Unsuppress the corresponding Hypothesis test.

Add a debugging utility that was helpful for hunting down the root
cause of that infinite loop, namely being able to print out
`self.evaluate` results of a dictionary that may have non-`Tensor`
values (such as the `parameters` dictionary of a `Distribution`).

Drive-by implement Bernoulli.cdf.

PiperOrigin-RevId: 328547724