Rename unnamed JD* RVs... again.
The goal of the previous renaming was allow namedtuple dtypes for
JointDistributions. It turned out to be insufficient, as namedtuples reject
fields that start with underscores.

Since the automatically generated name now can easily be written by a user by
accident, we also add a check for duplicate field names.

As with the previous rename, this is technically a breaking change, but the
scope should be extremely small so we're doing it without a deprecation cycle.

PiperOrigin-RevId: 328793442