Allow Hamiltonian Monte Carlo to use a momentum distribution for MCMC.
This adds two experimental classes:  `PreconditionedHamiltonianMonteCarlo` and `MultivariateNormalInverseScaleLinearOperator`.

PreconditionedHamiltonianMonteCarlo extends `tfp.mcmc.HamiltonianMonteCarlo`, and accepts a momentum distribution argument, which need not be an isotropic Gaussian. For example, if you suspect your posterior covariance is `S`, then using a kinetic energy function of `K(p) = 0.5 * (p.T S p)` may improve the performance of Hamiltonian Monte Carlo.

By providing a normal distribution with covariance `S^{-1}` as the momentum distribution, this is the kinetic energy function that will be used.

Additionally, a distribution, `MultivariateNormalInverseScaleLinearOperator`, is provided, whose implementation allows for careful control of performance, depending on the structure of a covariance estimate.

PiperOrigin-RevId: 334207371