PR #40778: Slightly improve performance of TensorShape methods in the common case
Imported from GitHub PR https://github.com/tensorflow/tensorflow/pull/40778

`TensorShape` is used a lot within the codebase and can end up on the hot path during eager execution. This PR simplifies some method calls to slightly improve performance for the case of fully defined shapes.

Most changes are related to the switch to a "Easier to ask for forgiveness than permission" (EAFP) pattern for accessing `TensorShape._dims` which is quite common in Python, but I would be happy to revert that if you prefer to still explicitely check for existance before access.
This PR improves performance of `.merge_elements()` by 50% and `.num_elements()` by 25%. `.rank`, `as_list` and `__add__` are 5-10% faster on my machine which probably doesn't really matter in practice.
Copybara import of the project:

--
8d93940cbee2d32ddafac4ba2d277f90cb60a09b by Lukas Geiger <lukas.geiger94@gmail.com>:

Slightly improve performance of TensorShape

PiperOrigin-RevId: 334589733