Add option for parallel filtering and sampling to tfd.LinearGaussianSSM.
Also remove some deprecated properties.

This change is a no-op for current users since `experimental_parallelize` is False by default. Ultimately it should probably be enabled by default; I've filed b/169178065 to track this.

This change slightly loosens the shape contract when returning a covariance matrix. Because the posterior covariance in an LGSSM depends only on the model parameters and observation mask, not the actual values of the observed time series, the existing sequential implementation can return covariance matrices that omit redundant sample_shape dimensions. The parallel implementation does not do this; the returned covariances always have full sample shape. Both versions satisfy the contract 'return a Tensor that broadcasts to the full sample + batch shape', so I've relaxed the documentation to describe this as the official contract.

PiperOrigin-RevId: 334864392