Inference Gym: Use a two-sample Z test when comparing to ground truth mean/stddev.
Not sure why we didn't do this in the first place. This'll go some way to
explain why these tests seemed so flaky (they were greatly underestimating the
standard error). Note that this has no effect on the stddev tests, as we
generally don't fill in the standard error of those estimates.

Also added some fudge factors for atols. We don't need them at the moment, but
it could be useful in cases where we don't have access to the ground truth
standard errors.

PiperOrigin-RevId: 334884988